@@ -12,8 +12,8 @@
     "cors": "^2.8.5",
     "express": "^4.17.3",
     "history": "^4.10.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.1.0",
+    "react-dom": "^18.1.0",
     "react-redux": "^7.2.6",
     "react-router": "^5.0.0",
     "react-router-dom": "^6.2.2",
